выделил константы в отдельный модуль
