+++ js/settings.js
@@ -1,0 +1,19 @@
+const PHOTOS_COUNT = 25;
+
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+const AVATARS_COUNT = 6;
+const COMMENTS_MIN = 5;
+const COMMENTS_MAX = 10;
+
+const NAMES = ['Иван', 'Николай', 'Михаил', 'Алексей', 'Владимир', 'Сергей', 'Артем'];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {PHOTOS_COUNT, LIKES_MIN, LIKES_MAX, AVATARS_COUNT, COMMENTS_MIN, COMMENTS_MAX, NAMES, MESSAGES};
